fix: admin course crud

+++ course_service/src/infrastructure/db/mongo/repositories/admin_course_repository.py
@@ -1,15 +1,25 @@
 from beanie.operators import Set
+from pymongo.errors import DuplicateKeyError
 
-from src.domain.dto.admin_course_dto import AdminCourseCreateDTO, AdminCourseUpdateDTO
+from src.domain.dto.admin_course_dto import (
+    AdminCourseCreateDTO,
+    AdminCourseUpdateDTO,
+)
 from src.domain.entities import CourseEntity
 from src.domain.repositories.admin_course_repository import IAdminCourseRepository
+from src.domain.exceptions.admin_course import CourseNotFoundError, CourseAlreadyExistsError
 
 
 class AdminCourseRepository(IAdminCourseRepository):
 
     async def create(self, dto: AdminCourseCreateDTO) -> CourseEntity:
-        course = CourseEntity(**dto.dict())
-        return await course.insert()
+        try:
+            course = CourseEntity(**dto.model_dump())
+            return await course.insert()
+        except DuplicateKeyError:
+            raise CourseAlreadyExistsError(
+                "Course with this title already exists"
+            )
 
     async def get(self, course_id: str) -> CourseEntity | None:
         return await CourseEntity.get(course_id)
@@ -18,13 +28,18 @@ class AdminCourseRepository(IAdminCourseRepository):
         return await CourseEntity.find_all().to_list()
 
     async def update(self, course_id: str, data: dict) -> CourseEntity:
-        await CourseEntity.find_one(
+        result = await CourseEntity.find_one(
             CourseEntity.id == course_id
-        ).update(
-            Set(data)
-        )
+        ).update(Set(data))
+
+        if result.matched_count == 0:
+            raise CourseNotFoundError(course_id)
 
         return await self.get(course_id)
 
-    async def delete(self, course: CourseEntity):
-        return await course.delete()
+    async def delete(self, course_id: str):
+        course = await self.get(course_id)
+        if not course:
+            raise CourseNotFoundError(course_id)
+
+        await course.delete()
